Improve Monaspace Disk Usage
At 35MB, I was fine leaving a few files on disk.

At 900+, I am not. There's a 24MB image on an M?

This change removes uninstalled files (as of version 1.200) and zeroes installed ones, so we can still use them for uninstall.

A better method of filtering may be required in the future.